requirements.txt
typer
rich
requests
beautifulsoup4
ollama
mmh3
shodan
python-whois
pyvis
crcmod
mmh3